Move sanitize URL to URLUtil and moved corresponding tests
